[front] Layout: dissolve AdminLayout into per-page archetypes (admin snap)
Phase 2 (2/3) of the layout system (design_docs/LAYOUT_SYSTEM.md §3.5, AP3
"Shadow shells"). Intentional visual change, reviewed as such:

- New AdminSubNavigation (front/components/navigation/) carries the admin
  sidebar sub-navigation registration only; RequireRoleLayout (front-spa)
  now mounts it instead of AdminLayout; AdminLayout is deleted
- The 16 in-shell admin pages declare their archetype:
  wide: MembersPage, AnalyticsPage, UsagePage, CreditsUsagePage
  centered: the 12 settings-style pages
- Bottom-padding parity (AdminLayout had pb-4 sm:pb-8; the shell has none):
  pb-8 added to MembersPage, AnalyticsPage, CreditsUsagePage roots
  (UsagePage already has pb-20; APIKeysPage keeps its h-12 spacer until
  Phase 3)

Visible diffs: max-w-6xl -> max-w-content (896px) on centered pages / full
width on wide pages; gutter breakpoint sm: -> md:; desktop top padding
32px -> 16px on centered pages.
