--- front/components/layouts/AdminLayout.tsx
@@ -1,54 +1,0 @@
-import { subNavigationAdmin } from "@app/components/navigation/config";
-import {
-  useSetContentWidth,
-  useSetSubNavigation,
-} from "@app/components/sparkle/AppLayoutContext";
-import {
-  useAuth,
-  useFeatureFlags,
-  useWorkspace,
-} from "@app/lib/auth/AuthContext";
-import { useAppRouter } from "@app/lib/platform";
-import { cn } from "@dust-tt/sparkle";
-import type { ReactElement } from "react";
-import { useMemo } from "react";
-
-interface AdminLayoutProps {
-  children: ReactElement;
-}
-
-export function AdminLayout({ children }: AdminLayoutProps) {
-  const owner = useWorkspace();
-  const { subscription } = useAuth();
-
-  const { featureFlags } = useFeatureFlags();
-
-  const router = useAppRouter();
-
-  const subNavigation = useMemo(
-    () =>
-      subNavigationAdmin({
-        owner,
-        currentRoute: router.pathname,
-        featureFlags,
-        subscription,
-      }),
-    [owner, router.pathname, featureFlags, subscription]
-  );
-
-  useSetSubNavigation(subNavigation);
-  // Temporary: AdminLayout renders its own max-w-6xl wrapper below, so the
-  // shell must stay hands-off. Removed when AdminLayout dissolves into
-  // per-page archetypes (LAYOUT_SYSTEM.md Phase 2, admin snap).
-  useSetContentWidth("full");
-
-  return (
-    <div
-      className={cn("flex h-full w-full flex-col items-center pt-4 sm:pt-8")}
-    >
-      <div className="flex w-full max-w-6xl grow flex-col px-4 sm:px-8 pb-4 sm:pb-8">
-        {children}
-      </div>
-    </div>
-  );
-}
