+++ front/components/pages/workspace/developers/APIKeysPage.tsx
@@ -1,3 +1,4 @@
+import { useSetContentWidth } from "@app/components/sparkle/AppLayoutContext";
 import { APIKeyCreationSheet } from "@app/components/workspace/api-keys/APIKeyCreationSheet";
 import { APIKeysList } from "@app/components/workspace/api-keys/APIKeysList";
 import { EditKeyCapDialog } from "@app/components/workspace/api-keys/EditKeyCapDialog";
@@ -239,6 +240,7 @@ export function APIKeys({ owner }: APIKeysProps) {
 }
 
 export function APIKeysPage() {
+  useSetContentWidth("centered");
   const owner = useWorkspace();
 
   return (
